.github/scripts/requirements.txt
PyYAML
pytablewriter